+++ provision/core/vvv/provision.sh
@@ -29,6 +29,9 @@ function vvv_register_packages() {
 
   VVV_PACKAGE_LIST+=(
     software-properties-common
+    
+    # Daily automatic security package upgrades
+    unattended-upgrades
 
     # other packages that come in handy
     subversion
@@ -39,9 +42,10 @@ function vvv_register_packages() {
     make
     vim
     colordiff
-    python3-pip
+    python3-pip # needed for shyaml
     python3-setuptools
     lftp
+    jq
 
     # ntp service to keep clock current
     ntp